Задачи: 1.Вывод всех данных продукта при запросе в постман в favorite,cart and order
