+++ src/main/java/com/example/marketplace_backend/Model/Order.java
@@ -70,6 +70,5 @@ public class Order {
 
     @Column(name = "order_number", unique = true, nullable = false)
     private Long orderNumber;
-
 }
