+++ src/main/java/com/example/marketplace_backend/controller/FavoriteController.java
@@ -1,12 +1,19 @@
 package com.example.marketplace_backend.controller;
 
 import com.example.marketplace_backend.DTO.Requests.models.FavoriteRequest;
+import com.example.marketplace_backend.DTO.Responses.models.BrandResponse;
+import com.example.marketplace_backend.DTO.Responses.models.FavoriteItemResponse;
 import com.example.marketplace_backend.DTO.Responses.models.FavoriteResponse;
+import com.example.marketplace_backend.Model.Brand;
 import com.example.marketplace_backend.Model.Favorite;
 import com.example.marketplace_backend.Model.Intermediate_objects.FavoriteItem;
+import com.example.marketplace_backend.Service.Impl.ConverterService;
 import com.example.marketplace_backend.Service.Impl.FavoriteService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,12 +25,12 @@ import java.util.List;
 @Tag(name = "Favorite API", description = "API для управления избранными товарами пользователей")
 public class FavoriteController {
     private final FavoriteService favoriteService;
-
+    private final ConverterService converterService;
 
     @PostMapping(value = "/add", consumes = {"multipart/form-data"})
     public ResponseEntity<FavoriteResponse> addItem(@ModelAttribute FavoriteRequest request) {
         Favorite favorite = favoriteService.addItemToFavorite(request.getProductId(), request.getQuantity());
-        return ResponseEntity.ok(favoriteService.convertToFavoriteResponse(favorite));
+        return ResponseEntity.ok(converterService.convertToFavoriteResponse(favorite));
     }
 
     @DeleteMapping("/remove")
@@ -41,11 +48,22 @@ public class FavoriteController {
     @GetMapping
     public ResponseEntity<FavoriteResponse> getFavorite() {
         Favorite favorite = favoriteService.getFavorite();
-        return ResponseEntity.ok(favoriteService.convertToFavoriteResponse(favorite));
+        return ResponseEntity.ok(converterService.convertToFavoriteResponse(favorite));
     }
 
+//    @GetMapping("/items")
+//    public ResponseEntity<List<FavoriteItem>> getFavoriteItems() {
+//        return favoriteService.getFavoriteItemsByUserId();
+//    }
     @GetMapping("/items")
-    public ResponseEntity<List<FavoriteItem>> getFavoriteItems() {
-        return favoriteService.getFavoriteItemsByUserId();
+    public ResponseEntity<Page<FavoriteItemResponse>> getPaginatedFavoriteItems(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size
+    ) {
+        Pageable pageable = PageRequest.of(page, size);
+        Page<FavoriteItem> favoriteItems = favoriteService.findAllItems(pageable);
+
+        Page<FavoriteItemResponse> responses = favoriteItems.map(converterService::convertToFavoriteItemResponse);
+        return ResponseEntity.ok(responses);
     }
 }
